Move app_util.py functions into app.py
